Reorganizing methods into worker.py, adding code comments
