+++ nexus/ui_connector/teams_activity_worker/src/teams_activity_worker/__init__.py
@@ -1,5 +1,5 @@
 """Microsoft Teams Temporal activity worker."""
 
-from .platform import TeamsActivities
+from .worker import TeamsActivities
 
 __all__ = ["TeamsActivities"]
