+++ nexus/ui_connector/teams_activity_worker/tests/test_platform.py
@@ -1,7 +1,5 @@
 import asyncio
-import runpy
 from dataclasses import dataclass
-from importlib.metadata import entry_points
 from typing import Any
 
 import pytest
@@ -19,10 +17,7 @@ from teams_activity_worker.platform import (
     MESSAGE_UPDATE_INTERVAL_SECONDS,
     STREAM_MODE_MESSAGE_UPDATE,
     STREAM_MODE_NATIVE,
-    Settings,
     TeamsPlatform,
-    _worker_task_queue,
-    main,
 )
 
 
@@ -110,27 +105,6 @@ def metadata(*, text: str = "", thread_id: str = "", service_url: str = "https:/
         service_url=service_url,
         channel_id="msteams",
     )
-
-
-def test_console_script_runs_platform_main() -> None:
-    entry_point = next(item for item in entry_points(group="console_scripts") if item.name == "teams-activity-worker")
-
-    assert entry_point.value == "teams_activity_worker.platform:main"
-    assert entry_point.load() is main
-
-
-def test_module_execution_runs_platform_main(monkeypatch: pytest.MonkeyPatch) -> None:
-    calls: list[Any] = []
-
-    def fake_run(coroutine: Any) -> None:
-        calls.append(coroutine)
-        coroutine.close()
-
-    monkeypatch.setattr(asyncio, "run", fake_run)
-    with pytest.warns(RuntimeWarning, match="found in sys.modules"):
-        runpy.run_module("teams_activity_worker.platform", run_name="__main__")
-
-    assert len(calls) == 1
 
 
 @pytest.mark.asyncio
@@ -277,38 +251,9 @@ async def test_update_message_replaces_approval_card(fixture) -> None:
 
 
 @pytest.mark.asyncio
-async def test_sdk_app_initializes_for_proactive_messaging() -> None:
-    platform = TeamsPlatform.from_settings(
-        Settings(
-            microsoft_tenant_id="tenant",
-            microsoft_app_id="app",
-            microsoft_app_password="secret",
-        ),
-        "teams-worker-1",
-    )
-
-    assert platform.app is not None
-    await platform.app.initialize()
-    activity_operations = platform._activities(metadata())
-    assert callable(activity_operations.create)
-    assert callable(activity_operations.reply)
-    assert callable(activity_operations.update)
-    await platform.app.stop()
-
-
-@pytest.mark.asyncio
 async def test_update_rejects_stream_owned_by_another_worker(fixture) -> None:
     platform, _, _ = fixture
     handle = StreamHandle("missing", "teams:conversation-1", STREAM_MODE_NATIVE, "another-worker")
 
     with pytest.raises(ValueError, match="not active"):
         await platform.update_stream(UpdateStream(metadata(), handle, "hello", "hello"))
-
-
-def test_worker_task_queues_are_unique_per_process() -> None:
-    first = _worker_task_queue("nexus-connector-teams")
-    second = _worker_task_queue("nexus-connector-teams")
-
-    assert first.startswith("nexus-connector-teams-stream-")
-    assert second.startswith("nexus-connector-teams-stream-")
-    assert first != second
